Prune out tests that are incomplete and/or of no further use. The newer end-to-end tests cover things like NR state actions and transitions.
